# src/main/resources/REST-API-DOC-swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8443'
basePath: /
...
paths:
  /cro/companies:
    get:
      tags:
        - Company Registrations Office REST API Interface
      summary: Get/search a list of all companies stored in application.
      description: Retrieve all companies details.
      operationId: searchCompaniesUsingGET
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: CompanyName
          in: query
          description: name of the company
          required: false
          type: string
        - name: id
          in: query
          description: searching for a company by id
          required: false
          type: integer
          format: int32
        - name: showall
          in: query
          description: Show companies with all stakeholders details.
          required: false
          type: boolean
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Company'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Company Registrations Office REST API Interface
      summary: Receive a company details as a request to be created on Company Registration Office(CRO).
      description: Receive a company details as a request to be created on Company Registration Office(CRO)
      operationId: registerCompanyUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/CompanyRegistrationRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Company'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/cro/companies/{companyId}':
    get:
      tags:
        - Company Registrations Office REST API Interface
      summary: Search a company based on the received company Id.
      description: Search a company based on the received company Id.
      operationId: getCompanyByIdUsingGET
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: companyId
          in: path
          description: companyId
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Company'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - Company Registrations Office REST API Interface
      summary: Receive a company details as an update request to be updated on Company Registration Office.
      description: Receive an updateRequest for update an existing company in store.
      operationId: updateCompanyUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: companyId
          in: path
          description: companyId
          required: true
          type: integer
          format: int32
        - in: body
          name: updateRequest
          description: updateRequest
          required: true
          schema:
            $ref: '#/definitions/CompanyRegistrationRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Company'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /cro/owners:
    post:
      tags:
        - Company Registrations Office REST API Interface
      summary: Receive a company Id and a list of stakeholders Id's as a request to be registered as beneficial owner of the company.
      description: Receive a company Id and a list of stakeholders Id's as a request to be registered as beneficial owner of the company.
      operationId: registerBeneficialOwnerUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - in: body
          name: req
          description: req
          required: true
          schema:
            $ref: '#/definitions/BeneficialRegisterRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Company'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
  BeneficialRegisterRequest:
    type: object
    required:
      - CompanyId
      - Stakeholders
    properties:
      CompanyId:
        type: integer
        format: int32
        description: Company Id
        allowEmptyValue: false
      Stakeholders:
        type: array
        description: A collection of the company stakeholders ID's
        allowEmptyValue: false
        items:
          type: integer
          format: int32
    title: BeneficialRegisterRequest
    description: Beneficial owner/stakeholder registration request resource representation
  Company:
    type: object
    required:
      - Stakeholders
    properties:
      Address:
        type: string
      City:
        type: string
      Country:
        type: string
      Email:
        type: string
      Id:
        type: integer
        format: int32
      Name:
        type: string
      PhoneNumber:
        type: string
      Stakeholders:
        type: array
        description: A collection of the returned company stakeholders
        allowEmptyValue: false
        items:
          $ref: '#/definitions/Stakeholder'
    title: Company
    description: Company resource representation
  CompanyRegistrationRequest:
    type: object
    properties:
      FirstName:
        type: string
      LastName:
        type: string
    title: CompanyRegistrationRequest
    description: Create Stakeholder registration request resource representation
...
  Stakeholder:
    type: object
    required:
      - Companies
    properties:
      Companies:
        type: array
        description: A list of companies details owned by this stakeholder
        allowEmptyValue: false
        items:
          $ref: '#/definitions/Company'
      FirstName:
        type: string
      Id:
        type: integer
        format: int32
      LastName:
        type: string
    title: Stakeholder
    description: Stakeholder resource representation
...